Reorganize codebase
